personal-finance-chatbot/requirements.txt
streamlit>=1.28.0
transformers>=4.30.0
torch>=2.0.0
huggingface_hub>=0.16.0
accelerate>=0.20.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
nltk>=3.8.0
matplotlib>=3.7.0
plotly>=5.15.0
requests>=2.31.0
scikit-learn>=1.3.0
google-generativeai>=0.3.0